+++ packages/bibbidi/lib/bibbidi/events/script/realm_destroyed.ex
@@ -11,4 +11,8 @@ defmodule Bibbidi.Events.Script.RealmDestroyed do
   """
 
   defstruct [:realm]
+
+  @doc "Returns the BiDi method name this event struct represents."
+  @spec method() :: String.t()
+  def method, do: "script.realmDestroyed"
 end